requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
requests>=2.31.0
pandas>=2.0.0
py-toon-format>=0.1.0
google-genai>=0.8.3